+++ src/Component/Facet/JsonTerms.php
@@ -51,11 +51,6 @@ class JsonTerms extends AbstractFacet implements JsonFacetInterface, FacetSetInt
     const METHOD_SMART = 'smart';
 
     /**
-     * @deprecated Use {@link SORT_COUNT_ASC} or {@link SORT_COUNT_DESC}
-     */
-    const SORT_COUNT = 'count';
-
-    /**
      * Sort buckets by 'count asc'.
      */
     const SORT_COUNT_ASC = 'count asc';
@@ -64,11 +59,6 @@ class JsonTerms extends AbstractFacet implements JsonFacetInterface, FacetSetInt
      * Sort buckets by 'count desc'.
      */
     const SORT_COUNT_DESC = 'count desc';
-
-    /**
-     * @deprecated Use {@link SORT_INDEX_ASC} or {@link SORT_INDEX_DESC}
-     */
-    const SORT_INDEX = 'index';
 
     /**
      * Sort buckets by 'index asc'.